just a backup

+++ lib/data_logger/provider/max_value_provider.dart
@@ -3,12 +3,12 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 // Class to hold both maxSpeed and maxRpm
 class MaxValues {
   final double maxSpeed;
-  final double maxRpm;
+  final int maxRpm;
 
   MaxValues({required this.maxSpeed, required this.maxRpm});
 
   // Create a copy of the current MaxValues with updated maxSpeed or maxRpm
-  MaxValues copyWith({double? maxSpeed, double? maxRpm}) {
+  MaxValues copyWith({double? maxSpeed, int? maxRpm}) {
     return MaxValues(
       maxSpeed: maxSpeed ?? this.maxSpeed,
       maxRpm: maxRpm ?? this.maxRpm,
@@ -25,7 +25,7 @@ class MaxValueProvider extends StateNotifier<MaxValues> {
   }
 
   // Update the maxRpm
-  void setMaxRpmValue(double value) {
+  void setMaxRpmValue(int value) {
     state = state.copyWith(maxRpm: value);
   }
 }
